Module Tezos_shell

module Block_directory : sig ... end

module Block_validator : sig ... end

This module is the main entry point to valide blocks and protocols.

module Block_validator_process : sig ... end

Block_validator_process is used to validate new blocks. This validation can be

module Bootstrap_pipeline : sig ... end

The bootstrap pipeline works as follows: 1. From a locator, it computes a list of subchains (identified by a Block_locator.step) to fetch. 2. A worker starts to fetch all the headers (top to bottom) from a subchain, starting with the top subchain. 3. A worker starts to download the list of operations by batch of blocks once a batch of headers is available. 4. A worker validates blocks one by one (bottom to top).

module Bootstrap_pipeline_event : sig ... end

module Chain_directory : sig ... end

module Chain_validator : sig ... end

module Chain_validator_event : sig ... end

module Clock_drift : sig ... end

module Config_directory : sig ... end

module Consensus_heuristic : sig ... end

module Distributed_db : sig ... end

Tezos Shell - High-level API for the Gossip network and local storage.

module Distributed_db_event : sig ... end

module Distributed_db_message : sig ... end

Tezos Shell - Network message for the gossip P2P protocol.

module Distributed_db_requester : sig ... end

This module creates FULL_REQUESTER modules for several resources. Protocols, operation, block_header, operation_hashes, operations.

module Injection_directory : sig ... end

module Monitor_directory : sig ... end

module Node : sig ... end

module Node_event : sig ... end

Declares logging events for node

module P2p_reader : sig ... end

A P2p_reader.t is a worker that answers the application messages (see Messages) received from a remote peer on a P2p.connection.

module Patch_context : sig ... end

module Peer_validator : sig ... end

module Prevalidation : sig ... end

A newly received block is validated by replaying locally the block creation, applying each operation and its finalization to ensure their consistency. This module is stateless and creates and manipulates the prevalidation_state.

module Prevalidator : sig ... end

Tezos Shell - Prevalidation of pending operations (a.k.a Mempool)

module Prevalidator_filters : sig ... end

module Protocol_directory : sig ... end

module Protocol_validator : sig ... end

module Protocol_validator_event : sig ... end

Declares logging events for protocol_validator

module Stat_directory : sig ... end

module Synchronisation_heuristic : sig ... end

module Validator : sig ... end

Tezos Shell - Main entry point of the validation scheduler.

module Validator_event : sig ... end

Declares logging events for validator

module Version_directory : sig ... end

module Worker_directory : sig ... end

module Worker_logger : sig ... end