Module Tezos_shell.Block_validator

type t

type limits = {

protocol_timeout : Tezos_base.Time.System.Span.t;

worker_limits : Tezos_base.Worker_types.limits;

}

val create : limits -> Distributed_db.t -> Block_validator_process.t -> start_testchain:bool -> t Tezos_base__TzPervasives.tzresult Lwt.t

create limits ddb bvp start_testchain creates a Block_validator.

limits contains various timeout limits.

ddb is used to commit a block on the storage and get the state of the chain for which the block is submitted to validation.

bvp is an instance of the Block_validator_process. bvp is a proxy between the shell and the validation part related to the economic protocol (See Block_validator_process).

start_testchain if set to true allows to run the testchain.

This function is not supposed to fail. It is implemented this way because of the interface implemented by the Worker module.

val validate : t -> ?⁠canceler:Lwt_canceler.t -> ?⁠peer:Tezos_base.P2p_peer.Id.t -> ?⁠notify_new_block:(Tezos_store.Store.Block.t -> unit) -> Distributed_db.chain_db -> Tezos_crypto.Block_hash.t -> Tezos_base.Block_header.t -> Tezos_base.Operation.t list list -> unit Tezos_base__TzPervasives.tzresult Lwt.t

validate validator ddb hash header ops validates a block header ops of hash hash. It is a no-op in the following cases:

If the block has already been validated.

If the block level is before the savepoint

Otherwise it calls the Block_validator_process process associated to the current validator.

canceler is trigerred when the validation of a block fails.

peer is the peer which sent the block.

If the validation succeeded it processes as follows:

1. The ddb commits the block on the storage.

2. If the next block requires a switch of protocol, it tries to fetch and precompile the next protocol.

3. Call notify_new_block with the committed block.

An error is raised if the validation failed or if the block was already known as invalid. However, if the first validation attempt failed because the protocol was missing, it tries to fetch and download the protocol before trying to validate the block a second time.

val fetch_and_compile_protocol : t -> ?⁠peer:Tezos_base.P2p_peer.Id.t -> ?⁠timeout:Tezos_base.Time.System.Span.t -> Tezos_crypto.Protocol_hash.t -> Tezos_protocol_updater.Registered_protocol.t Tezos_base__TzPervasives.tzresult Lwt.t

val shutdown : t -> unit Lwt.t

val running_worker : unit -> t

val status : t -> Tezos_base.Worker_types.worker_status

val pending_requests : t -> (Tezos_base.Time.System.t * Tezos_shell_services.Block_validator_worker_state.Request.view) list

val current_request : t -> (Tezos_base.Time.System.t * Tezos_base.Time.System.t * Tezos_shell_services.Block_validator_worker_state.Request.view) option

val last_events : t -> (Tezos_base__TzPervasives.Internal_event.level * Tezos_shell_services.Block_validator_worker_state.Event.t list) list