Module Tezos_shell.Block_validator_process

type validator_environment = {

user_activated_upgrades : Tezos_base.User_activated.upgrades;
user activated upgrades

user_activated_protocol_overrides : Tezos_base.User_activated.protocol_overrides;
user activated protocol overrides

}

type validator_kind =

| Internal : Tezos_store.Store.Chain.chain_store -> validator_kind

| External : {

genesis : Tezos_base.Genesis.t;

data_dir : string;

context_root : string;

protocol_root : string;

process_path : string;

sandbox_parameters : Tezos_base__TzPervasives.Data_encoding.json option;

} -> validator_kind

For performances reasons, it may be interesting to use another processus (from the OS) to validate blocks (External). However, in that case, only one processus has a write access to the context. Currently informations are exchanged via the file system.

type t

Internal representation of the block validator process

val init : validator_environment -> validator_kind -> t Tezos_base__TzPervasives.tzresult Lwt.t

val close : t -> unit Lwt.t

close vp closes the given vp. In the case of an External validator process, we first ask the validator to shutdown. If it is still running after 5 seconds, we notice that the block validation process is unresponsive and we force its termination (using a registered Lwt_exit.clean_up_callback).

val restore_context_integrity : t -> int option Tezos_base__TzPervasives.tzresult Lwt.t

val apply_block : t -> Tezos_store.Store.chain_store -> predecessor:Tezos_store.Store.Block.t -> Tezos_base.Block_header.t -> Tezos_base.Operation.t list list -> Tezos_validation.Block_validation.result Tezos_base__TzPervasives.tzresult Lwt.t

apply_block bvp predecessor header os checks the liveness of the operations and then call Block_validation.apply

val commit_genesis : t -> chain_id:Tezos_crypto.Chain_id.t -> Tezos_crypto.Context_hash.t Tezos_base__TzPervasives.tzresult Lwt.t

val init_test_chain : t -> Tezos_store.Store.Block.t -> Tezos_base.Block_header.t Tezos_base__TzPervasives.tzresult Lwt.t

init_test_chain must only be called on a forking block.