Module Distributed_db.Request

val current_branch : chain_db -> ?⁠peer:Tezos_base.P2p_peer.Id.t -> unit -> unit

current_branch chain_db ?peer () sends a Get_current_branch chain_id message to peer, or if peer isn't specified, to all known active peers for this chain. chain_id is the identifier for chain_db. Expected answer is a Current_branch message.

val current_head : chain_db -> ?⁠peer:Tezos_base.P2p_peer.Id.t -> unit -> unit

current_header chain_db ?peer () sends a Get_Current_head chain_id to a given peer, or to all known active peers for this chain. chain_id is the identifier for chain_db. Expected answer is a Get_current_head message