Module Tezos_shell.Prevalidation

A newly received block is validated by replaying locally the block creation, applying each operation and its finalization to ensure their consistency. This module is stateless and creates and manipulates the prevalidation_state.

module type T = sig ... end

module Make (Proto : Tezos_protocol_updater.Registered_protocol.T) : T with module Proto = Proto

val preapply : Tezos_store.Store.chain_store -> user_activated_upgrades:Tezos_base.User_activated.upgrades -> user_activated_protocol_overrides:Tezos_base.User_activated.protocol_overrides -> predecessor:Tezos_store.Store.Block.t -> timestamp:Tezos_base.Time.Protocol.t -> protocol_data:Stdlib.Bytes.t -> Tezos_base.Operation.t list list -> (Tezos_base.Block_header.shell_header * Tezos_error_monad.TzCore.error Tezos_base.Preapply_result.t list, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Pre-apply creates a new block and returns it.