Module Tezos_shell.Synchronisation_heuristic

Synchronisation heuristic

type status = Tezos_shell_services.Chain_validator_worker_state.Event.synchronisation_status =

| Synchronised of {

is_chain_stuck : bool;

}

| Not_synchronised

type t

internal state of the bootstrap_heuristic

val update : t -> (Tezos_base.Time.Protocol.t * Tezos_base.P2p_peer.Id.t) -> unit

update t (timestamp, peer) updates t according the heuristic above. The timestamp should come from a valid block, and the peer should be the one who sent the block.

val get_status : t -> status

get_status t gives the current status according of the bootstrap heuristic described above.

val create : threshold:int -> latency:int -> t

create ~threshold ~sync_latency initializes the heuristic with these two parameters