Module Tezos_shell__Block_validator_process

Block_validator_process is used to validate new blocks. This validation can be

internal: the same processus is used to run the node and to validate blocks

external: another processus is used to validate blocks This module also ensures the liveness of the operations (see Block_validation:check_liveness).