Module Tezos_shell__Bootstrap_pipeline

The bootstrap pipeline works as follows: 1. From a locator, it computes a list of subchains (identified by a Block_locator.step) to fetch. 2. A worker starts to fetch all the headers (top to bottom) from a subchain, starting with the top subchain. 3. A worker starts to download the list of operations by batch of blocks once a batch of headers is available. 4. A worker validates blocks one by one (bottom to top).

These three workers work concurrently.