Module Tezos_shell.Bootstrap_pipeline

type t

val create : ?⁠notify_new_block:(Tezos_store.Store.Block.t -> unit) -> block_header_timeout:Tezos_base.Time.System.Span.t -> block_operations_timeout:Tezos_base.Time.System.Span.t -> Block_validator.t -> Tezos_base.P2p_peer.Id.t -> Distributed_db.chain_db -> Tezos_base.Block_locator.t -> t

create ?notify_new_block ~block_header_timeout ~block_operations_timeout validator peer ddb locator initializes a bootstrap_pipeline for a locator sent by peer. It uses the Distributed_db to fetch the data needed and validator to validate blocks one by one with the function Block_validator.validate.

Moreover:

notify_new_block is a called everytime a valid block is received.

block_header_timeout is the timeout to fetch a block_header.

block_operations_timeout is the timeout to fetch a block operation.

If a timeout is triggered, the whole bootstrap_pipeline is canceled.

val wait : t -> unit Tezos_base__TzPervasives.tzresult Lwt.t

val cancel : t -> unit Lwt.t

val length : t -> Tezos_shell_services.Peer_validator_worker_state.Worker_state.pipeline_length

length pipeline returns the number of the headers and blocks fetched

val length_zero : Tezos_shell_services.Peer_validator_worker_state.Worker_state.pipeline_length